@@ -28,43 +65,7 @@
     "leveldb",
     "level"
   ],
-  "main": "leveldown.js",
-  "typings": "leveldown.d.ts",
-  "dependencies": {
-    "abstract-leveldown": "~4.0.0",
-    "bindings": "~1.3.0",
-    "fast-future": "~1.0.2",
-    "nan": "~2.9.1",
-    "prebuild-install": "^2.1.0"
-  },
-  "devDependencies": {
-    "async": "^2.0.1",
-    "delayed": "~1.0.1",
-    "du": "~0.1.0",
-    "iota-array": "~1.0.0",
-    "lexicographic-integer": "~1.1.0",
-    "mkfiletree": "~1.0.1",
-    "monotonic-timestamp": "~0.0.8",
-    "node-uuid": "~1.4.3",
-    "optimist": "~0.6.1",
-    "prebuild": "^7.0.0",
-    "prebuild-ci": "^2.0.0",
-    "readfiletree": "~0.0.1",
-    "rimraf": "^2.6.1",
-    "slump": "~2.0.0",
-    "tap-spec": "^4.1.1",
-    "tape": "^4.5.1",
-    "verify-travis-appveyor": "^2.0.2"
-  },
-  "scripts": {
-    "install": "prebuild-install || node-gyp rebuild",
-    "test": "verify-travis-appveyor && tape test/*-test.js | tap-spec && prebuild-ci",
-    "rebuild": "prebuild --compile",
-    "prebuild": "prebuild -t 8.1.2 -t 6.11.0 -t 4.8.3 -t 7.10.0 --strip --verbose"
-  },
-  "license": "MIT",
-  "gypfile": true,
   "engines": {
-    "node": ">=4"
+    "node": ">=8.6.0"
   }
 }
