@@ -1,16 +1,17 @@
 const BUFFERS = false
 
-var leveldown = require('..')
-  , crypto    = require('crypto')
-  , putCount  = 0
-  , getCount  = 0
-  , rssBase
-  , db
+const testCommon = require('./common')
+const crypto = require('crypto')
+
+let putCount = 0
+let getCount = 0
+let rssBase
+let db
 
 function run () {
   var key = 'long key to test memory usage ' + String(Math.floor(Math.random() * 10000000))
 
-  if (BUFFERS) key = new Buffer(key)
+  if (BUFFERS) key = Buffer.from(key)
 
   db.get(key, function (err, value) {
     getCount++
@@ -29,27 +30,18 @@ function run () {
   })
 
   if (getCount % 1000 === 0) {
-    if (typeof gc != 'undefined')
-      gc()
-    console.log(
-        'getCount ='
-      , getCount
-      , ', putCount = '
-      , putCount
-      , ', rss ='
-      , Math.round(process.memoryUsage().rss / rssBase * 100) + '%'
-      , Math.round(process.memoryUsage().rss / 1024 / 1024) + 'M'
-      , JSON.stringify([0,1,2,3,4,5,6].map(function (l) {
-          return db.getProperty('leveldb.num-files-at-level' + l)
-        }))
-    )
+    if (typeof global.gc !== 'undefined') global.gc()
+    console.log('getCount =', getCount, ', putCount = ', putCount, ', rss =',
+      Math.round(process.memoryUsage().rss / rssBase * 100) + '%',
+      Math.round(process.memoryUsage().rss / 1024 / 1024) + 'M',
+      JSON.stringify([0, 1, 2, 3, 4, 5, 6].map(function (l) {
+        return db.getProperty('rocksdb.num-files-at-level' + l)
+      })))
   }
 }
 
-leveldown.destroy('./leakydb', function () {
-  db = leveldown('./leakydb')
-  db.open({ xcacheSize: 0, xmaxOpenFiles: 10 }, function () {
-    rssBase = process.memoryUsage().rss
-    run()
-  })
+db = testCommon.factory()
+db.open(function () {
+  rssBase = process.memoryUsage().rss
+  run()
 })
