@@ -1,35 +1,33 @@
-const test         = require('tape')
-    , testCommon   = require('abstract-leveldown/testCommon')
-    , fs           = require('fs')
-    , path         = require('path')
-    , mkfiletree   = require('mkfiletree')
-    , readfiletree = require('readfiletree')
-    , rimraf       = require('rimraf')
-    , leveldown    = require('../')
-    , makeTest     = require('./make')
+const test = require('tape')
+const tempy = require('tempy')
+const fs = require('fs')
+const path = require('path')
+const mkfiletree = require('mkfiletree')
+const readfiletree = require('readfiletree')
+const rimraf = require('rimraf')
+const leveldown = require('..')
+const makeTest = require('./make')
 
 test('test argument-less destroy() throws', function (t) {
-  t.throws(
-      leveldown.destroy
-    , { name: 'Error', message: 'destroy() requires `location` and `callback` arguments' }
-    , 'no-arg destroy() throws'
-  )
+  t.throws(leveldown.destroy, {
+    name: 'Error',
+    message: 'destroy() requires `location` and `callback` arguments'
+  }, 'no-arg destroy() throws')
   t.end()
 })
 
 test('test callback-less, 1-arg, destroy() throws', function (t) {
-  t.throws(
-      leveldown.destroy.bind(null, 'foo')
-    , { name: 'Error', message: 'destroy() requires `location` and `callback` arguments' }
-    , 'callback-less, 1-arg destroy() throws'
-  )
+  t.throws(leveldown.destroy.bind(null, 'foo'), {
+    name: 'Error',
+    message: 'destroy() requires `location` and `callback` arguments'
+  }, 'callback-less, 1-arg destroy() throws')
   t.end()
 })
 
 test('test destroy non-existent directory', function (t) {
   t.plan(4)
 
-  var location = testCommon.location()
+  var location = tempy.directory()
   var parent = path.dirname(location)
 
   // For symmetry with the opposite test below.
@@ -37,10 +35,10 @@ test('test destroy non-existent directory', function (t) {
 
   // Cleanup to avoid conflicts with other tests
   rimraf(location, { glob: false }, function (err) {
-    t.ifError(err, 'no rimraf error')
+    t.ifError(err, 'no error from rimraf()')
 
-    leveldown.destroy(location, function () {
-      t.is(arguments.length, 0, 'no arguments returned on callback')
+    leveldown.destroy(location, function (err) {
+      t.error(err, 'no error')
 
       // Assert that destroy() didn't inadvertently create the directory.
       // Or if it did, that it was at least cleaned up afterwards.
@@ -50,39 +48,38 @@ test('test destroy non-existent directory', function (t) {
 })
 
 test('test destroy non-existent parent directory', function (t) {
-  t.plan(4)
+  t.plan(3)
 
   var location = '/1/2/3/4'
   var parent = path.dirname(location)
 
   t.notOk(fs.existsSync(parent), 'parent does not exist before')
 
-  leveldown.destroy(location, function () {
-    // This behavior differs from leveldown, which is silent.
-    t.is(arguments.length, 1, 'error object returned on callback')
-    t.ok(/.*IO error.*\/1\/2\/3\/4\/LOCK.*/.test(arguments[0]), 'got IO error')
+  leveldown.destroy(location, function (err) {
+    // This behavior differs from LevelDB, which is silent.
+    t.ok(/.*IO error.*\/1\/2\/3\/4\/LOCK.*/.test(err), 'got IO error')
     t.notOk(fs.existsSync(location), 'directory does not exist after')
   })
 })
 
 test('test destroy non leveldb directory', function (t) {
   var tree = {
-      'foo': 'FOO'
-    , 'bar': { 'one': 'ONE', 'two': 'TWO', 'three': 'THREE' }
+    foo: 'FOO',
+    bar: { 'one': 'ONE', 'two': 'TWO', 'three': 'THREE' }
   }
 
   mkfiletree.makeTemp('destroy-test', tree, function (err, dir) {
-    t.ifError(err, 'no close error')
+    t.ifError(err, 'no error from makeTemp()')
 
     leveldown.destroy(dir, function (err) {
-      t.ifError(err, 'no destroy error')
+      t.ifError(err, 'no error from destroy()')
 
       readfiletree(dir, function (err, actual) {
-        t.ifError(err, 'no read error')
+        t.ifError(err, 'no error from readfiletree()')
         t.deepEqual(actual, tree, 'directory remains untouched')
 
         mkfiletree.cleanUp(function (err) {
-          t.ifError(err, 'no cleanup error')
+          t.ifError(err, 'no error from cleanup()')
           t.end()
         })
       })
@@ -90,33 +87,35 @@ test('test destroy non leveldb directory', function (t) {
   })
 })
 
-makeTest('test destroy() cleans and removes leveldb-only dir', function (db, t, done, location) {
+makeTest('test destroy() cleans and removes leveldb-only dir', function (db, t, done) {
+  var location = db.location
   db.close(function (err) {
-    t.ifError(err, 'no close error')
+    t.ifError(err, 'no error from close()')
 
     leveldown.destroy(location, function (err) {
-      t.ifError(err, 'no destroy error')
+      t.ifError(err, 'no error from destroy()')
       t.notOk(fs.existsSync(location), 'directory completely removed')
 
-      done(false)
+      done(null, false)
     })
   })
 })
 
-makeTest('test destroy() cleans and removes only leveldb parts of a dir', function (db, t, done, location) {
+makeTest('test destroy() cleans and removes only leveldb parts of a dir', function (db, t, done) {
+  var location = db.location
   fs.writeFileSync(path.join(location, 'foo'), 'FOO')
 
   db.close(function (err) {
-    t.ifError(err, 'no close error')
+    t.ifError(err, 'no error from close()')
 
     leveldown.destroy(location, function (err) {
-      t.ifError(err, 'no destroy error')
+      t.ifError(err, 'no error from destroy()')
 
       readfiletree(location, function (err, tree) {
-        t.ifError(err, 'no read error')
+        t.ifError(err, 'no error from readfiletree()')
         t.deepEqual(tree, { 'foo': 'FOO' }, 'non-leveldb files left intact')
 
-        done(false)
+        done(null, false)
       })
     })
   })
