@@ -1,23 +1,24 @@
-const test          = require('tape')
-    , testCommon    = require('abstract-leveldown/testCommon')
-    , leveldown     = require('../')
-    , child_process = require('child_process') 
+const test = require('tape')
+const testCommon = require('./common')
+const fork = require('child_process').fork
+const path = require('path')
 
-var db
-  , sourceData = (function () {
-      var d = []
-        , i = 0
-        , k
-      for (; i <  100000; i++) {
-        k = (i < 10 ? '0' : '') + i
-        d.push({
-            type  : 'put'
-          , key   : k
-          , value : Math.random()
-        })
-      }
-      return d
-    }())
+let db
+
+const sourceData = (function () {
+  var d = []
+  var i = 0
+  var k
+  for (; i < 100000; i++) {
+    k = (i < 10 ? '0' : '') + i
+    d.push({
+      type: 'put',
+      key: k,
+      value: Math.random()
+    })
+  }
+  return d
+}())
 
 test('setUp common', testCommon.setUp)
 
@@ -25,11 +26,11 @@ test('try to create an iterator with a blown stack', function (t) {
   // Reducing the stack size down from the default 984 for the child node
   // process makes it easier to trigger the bug condition. But making it too low
   // causes the child process to die for other reasons.
-  var opts  = { execArgv: [ '--stack-size=128' ] }
-    , child = child_process.fork(__dirname + '/stack-blower.js', [ 'run' ], opts)
+  var opts = { execArgv: [ '--stack-size=128' ] }
+  var child = fork(path.join(__dirname, 'stack-blower.js'), [ 'run' ], opts)
 
   t.plan(2)
-  
+
   child.on('message', function (m) {
     t.ok(true, m)
     child.disconnect()
@@ -41,7 +42,7 @@ test('try to create an iterator with a blown stack', function (t) {
 })
 
 test('setUp db', function (t) {
-  db = leveldown(testCommon.location())
+  db = testCommon.factory()
   db.open(function (err) {
     t.error(err)
     db.batch(sourceData, t.end.bind(t))
@@ -50,19 +51,17 @@ test('setUp db', function (t) {
 
 test('iterate over a large iterator with a large watermark', function (t) {
   var iterator = db.iterator({
-        highWaterMark: 10000000
-    })
-    , count = 0
-    , read = function () {
-        iterator.next(function () {
-          count++
-
-          if (!arguments.length)
-            t.end()
-          else
-            read()
-        })
+    highWaterMark: 10000000
+  })
+  var read = function () {
+    iterator.next(function () {
+      if (!arguments.length) {
+        t.end()
+      } else {
+        read()
       }
+    })
+  }
 
   read()
 })
