@@ -1,13 +1,11 @@
-/* Not sure if this is needed/supported with RocksDB. It looks like it might be
-   a small floated patch on leveldown.
 const test = require('tape')
-    , path = require('path')
-    , fs   = require('fs')
+const path = require('path')
+const fs = require('fs')
 
-test('test port-libuv is being used', function (t) {
+test.skip('test port-libuv is being used', function (t) {
   var version = fs.readFileSync(path.join(__dirname, '../deps/leveldb/leveldb.gyp'), 'utf8')
-                  .match(/'ldbversion': '([^']+)'/)[1]
-    , porth
+    .match(/"ldbversion": "([^"]+)"/)[1]
+  var porth
 
   t.ok(version, 'matched current leveldb version')
 
@@ -17,4 +15,3 @@ test('test port-libuv is being used', function (t) {
 
   t.end()
 })
-*/
