@@ -1,26 +1,21 @@
-const test         = require('tape')
-    , fs           = require('fs')
-    , path         = require('path')
-    , mkfiletree   = require('mkfiletree')
-    , readfiletree = require('readfiletree')
-    , leveldown    = require('../')
-    , makeTest     = require('./make')
+const test = require('tape')
+const fs = require('fs')
+const leveldown = require('..')
+const makeTest = require('./make')
 
 test('test argument-less repair() throws', function (t) {
-  t.throws(
-      leveldown.repair
-    , { name: 'Error', message: 'repair() requires `location` and `callback` arguments' }
-    , 'no-arg repair() throws'
-  )
+  t.throws(leveldown.repair, {
+    name: 'Error',
+    message: 'repair() requires `location` and `callback` arguments'
+  }, 'no-arg repair() throws')
   t.end()
 })
 
 test('test callback-less, 1-arg, repair() throws', function (t) {
-  t.throws(
-      leveldown.repair.bind(null, 'foo')
-    , { name: 'Error', message: 'repair() requires `location` and `callback` arguments' }
-    , 'callback-less, 1-arg repair() throws'
-  )
+  t.throws(leveldown.repair.bind(null, 'foo'), {
+    name: 'Error',
+    message: 'repair() requires `location` and `callback` arguments'
+  }, 'callback-less, 1-arg repair() throws')
   t.end()
 })
 
@@ -32,17 +27,24 @@ test('test repair non-existent directory returns error', function (t) {
 })
 
 // a proxy indicator that RepairDB is being called and doing its thing
-makeTest('test repair() compacts', function (db, t, done, location) {
+makeTest('test repair() compacts', function (db, t, done) {
+  var location = db.location
+
   db.close(function (err) {
-    t.notOk(err, 'no error')
+    t.ifError(err, 'no error from close()')
+
     var files = fs.readdirSync(location)
     t.ok(files.some(function (f) { return (/\.log$/).test(f) }), 'directory contains log file(s)')
     t.notOk(files.some(function (f) { return (/\.sst$/).test(f) }), 'directory does not contain sst file(s)')
-    leveldown.repair(location, function () {
+
+    leveldown.repair(location, function (err) {
+      t.ifError(err, 'no error from repair()')
+
       files = fs.readdirSync(location)
       t.notOk(files.some(function (f) { return (/\.log$/).test(f) }), 'directory does not contain log file(s)')
       t.ok(files.some(function (f) { return (/\.sst$/).test(f) }), 'directory contains sst file(s)')
-      done(false)
+
+      done(null, false)
     })
   })
 })
